requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
overrides==7.7.0
pyarr==5.2.0
python-dotenv==1.0.1
requests==2.32.3
types-requests==2.32.0.20240622
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
